move cache defaults
